+++ core/templates/dev/head/pages/exploration_editor/exploration_editor.html
@@ -84,7 +84,7 @@
       </improvements-tab>
 
       <div ng-show="getActiveTabName() === 'settings'">
-        {% include 'pages/exploration_editor/settings_tab/settings_tab.html' %}
+        <settings-tab></settings-tab>
       </div>
 
       <div ng-show="getActiveTabName() === 'history'">
